Move and rename DCEL functionality
This PR doesn't modify and behaviour. It consists entirely of renames
and code moves. These renames and moves were saved up from previous PRs,
where moves and renames would have obscured the important parts of the
changes.
